// src/types/api/user.ts
// jsonplaceholderから取ってくる user情報をそのままコピー、貼り付け → json形式を修正 → 1つづつ型を定義

export type User = {
  id: number;
  name: string;
  username: string;
  email: string;
  address: {
    street: string;
    suite: string;
    city: string;
    zipcode: string;
    geo: {
      lat: string;
      lng: string;
    };
  };
  phone: string;
  website: string;
  company: {
    name: string;
    catchPhrase: string;
    bs: string;
  };
};